Cambios hechos en clase a Rama Development

+++ demo/src/main/java/com/example/demo/repository/ClienteRepository.java
@@ -12,5 +12,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 public interface ClienteRepository extends JpaRepository<Cliente, Long>{
 
-    public Cliente findByCedula(Integer cedula);
+    public Cliente findByCedula(Long cedula);
 }
